+++ tests/cases/test_integration/test_basic.py
@@ -1,0 +1,7 @@
+
+
+def test_defaults(base_app):
+    '''Load defaults config for fullauth'''
+
+    # Config should get created based on fullauth defaults
+    assert 'fullauth' in base_app.config.registry['config']